backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
django-cors-headers>=4.3
psycopg[binary]>=3.1
cloudinary>=1.36
# pip install Pillow
# razorpay
# pip install celery redis django-celery-beat
